feat: restructure error handling by creating a dedicated errors module and refactoring related imports feat: add blog components and utilities for fetching and displaying blog posts feat: implement custom MDX components for enhanced markdown rendering
